requirements.txt
psycopg2
sqlalchemy
sqlalchemy_orm
flask
python-dotenv